+++ pytorch_lightning/trainer/trainer.py
@@ -83,7 +83,7 @@ from pytorch_lightning.utilities.argparse import (
 from pytorch_lightning.utilities.auto_restart import _detect_fault_tolerant_env_to_enum
 from pytorch_lightning.utilities.cloud_io import get_filesystem
 from pytorch_lightning.utilities.distributed import distributed_available
-from pytorch_lightning.utilities.enums import FaultTolerantTrainingMode
+from pytorch_lightning.utilities.enums import _FaultTolerantTrainingMode
 from pytorch_lightning.utilities.exceptions import ExitGracefullyException, MisconfigurationException
 from pytorch_lightning.utilities.imports import _fault_tolerant_training
 from pytorch_lightning.utilities.meta import is_on_meta_device, materialize_module
@@ -473,7 +473,7 @@ class Trainer(
         self._lightning_optimizers = None
 
         # detect the fault tolerant flag
-        self._fault_tolerant_mode: FaultTolerantTrainingMode = _detect_fault_tolerant_env_to_enum()
+        self._fault_tolerant_mode: _FaultTolerantTrainingMode = _detect_fault_tolerant_env_to_enum()
 
         # .validate() and .test() set this when they load a checkpoint
         self.validated_ckpt_path: Optional[str] = None